add expiration

+++ hola-cafe-management-system/src/utils/formatter.tsx
@@ -1,5 +1,5 @@
 export function toTitleCase(str: string) {
-  if (!str) return '';
+  if (!str) return "";
   return str.replace(
     /\w\S*/g,
     (txt) => txt.charAt(0).toUpperCase() + txt.substr(1).toLowerCase()
@@ -9,7 +9,7 @@ export function toTitleCase(str: string) {
 export function dateFormatter(dateStr: string) {
   const date = new Date(dateStr);
   return new Intl.DateTimeFormat("en-US", {
-    month: "long",
+    month: "short",
     day: "numeric",
     year: "numeric",
   }).format(date);
